+++ src/components/entryScreen.js
