+++ src/navigation/navmodal.js
@@ -13,7 +13,7 @@ import ExploreScreen from '../components/explorescreen.js';
 import GalleryScreen from '../components/galleryscreen.js';
 import ProfileScreen from '../components/profilescreen.js';
 import FriendsScreen from '../components/friendsscreen.js';
-import LoginScreen from '../components/loginScreen.js';
+import EntryScreen from '../components/entryScreen.js';
 import Signup from '../components/signup.js';
 import Login from '../components/login.js';
 
@@ -52,7 +52,7 @@ function MyNav() {
     return (
         <NavigationContainer>
             <Stack.Navigator screenOptions={{headerShown: false}}>
-                <Stack.Screen name="Entry" component={LoginScreen} />
+                <Stack.Screen name="Entry" component={EntryScreen} />
                 <Stack.Screen name="Drawer" component={MyDrawer} />
                 <Stack.Screen name="Signup" component={Signup} />
                 <Stack.Screen name="Login" component={Login} />
